修改明细的显示问题 修改我的课程代付款 支付问题

+++ tongzijunyh/src/com/bm/base/adapter/MyAccountBalanceAdapter.java
@@ -62,8 +62,8 @@ public class MyAccountBalanceAdapter  extends BaseAd<Order>{
 				itemView.tv_accoundTitle.setText("【评价赠送】"+getNullData(entity.integralName));//名称
 			}
 		}
-		
-		if (Integer.parseInt(entity.couponMoney)>0) {
+
+		if (Float.parseFloat(entity.couponMoney)>0) {
 			itemView.tv_word.setVisibility(View.VISIBLE);
 			itemView.tv_couponMoney.setVisibility(View.VISIBLE);
 			itemView.tv_couponMoney.setText("-"+entity.couponMoney);
